refactor(`Settings`): update import locations.
