+++ src/interface/index.ts
@@ -1,12 +1,7 @@
 // Length.
-export type { Length } from './length.interface';
-export type { LengthOptions } from './length-options.interface';
-export type { LengthSetting } from './length-setting.interface';
-export type { LengthSettings } from './length-settings.interface';
+export type { Length, LengthOptions, LengthSetting, LengthSettings } from './length';
 // Pattern.
-export type { PatternOptions } from './pattern-options.interface';
-export type { PatternSetting } from './pattern-setting.interface';
-export type { PatternSettings } from './pattern-settings.interface';
+export type { PatternOptions, PatternSetting, PatternSettings } from './pattern';
 // Settings.
 export type { Settings } from './settings.interface';
 export type { ValueSetting } from './value-setting.interface';
